// backend/src/enums/region-enum.ts
export enum RegionEnum {
    VINNYTSIA = "Вінницька",
    VOLYN = "Волинська",
    DNIPROPETROVSK = "Дніпропетровська",
    DONETSK = "Донецька",
    ZHYTOMYR = "Житомирська",
    ZAKARPATTIA = "Закарпатська",
    ZAPORIZHZHIA = "Запорізька",
    IVANO_FRANKIVSK = "Івано-Франківська",
    KYIV = "Київська",
    KIROVOHRAD = "Кіровоградська",
    LUHANSK = "Луганська",
    LVIV = "Львівська",
    MYKOLAIV = "Миколаївська",
    ODESA = "Одеська",
    POLTAVA = "Полтавська",
    RIVNE = "Рівненська",
    SUMY = "Сумська",
    TERNOPIL = "Тернопільська",
    KHARKIV = "Харківська",
    KHERSON = "Херсонська",
    KHMELNYTSKYI = "Хмельницька",
    CHERKASY = "Черкаська",
    CHERNIVTSI = "Чернівецька",
    CHERNIHIV = "Чернігівська",
    CRIMEA = "АР Крим",
}